+++ src/core/constraint.jl
@@ -102,3 +102,96 @@ function constraint_block_isolation(pm::AbstractSwitchModels, nw::Int; relax::Bo
         JuMP.@constraint(pm.model, z_block <= n_gen + n_strg + n_neg_loads + sum(var(pm, nw, :switch_state, s) for s in ids(pm, nw, :block_switches) if s in ids(pm, nw, :switch_dispatchable)))
     end
 end
+
+
+"""
+    constraint_radial_topology(pm::AbstractUnbalancedPowerModel, nw::Int; relax::Bool=false)
+
+Constraint to enforce a radial topology
+# doi: 10.1109/TSG.2020.2985087
+"""
+function constraint_radial_topology(pm::AbstractSwitchModels, nw::Int; relax::Bool=false)
+    # doi: 10.1109/TSG.2020.2985087
+    var(pm, nw)[:f] = Dict{Tuple{Int,Int,Int},JuMP.VariableRef}()
+    var(pm, nw)[:lambda] = Dict{Tuple{Int,Int},JuMP.VariableRef}()
+    var(pm, nw)[:beta] = Dict{Tuple{Int,Int},JuMP.VariableRef}()
+    var(pm, nw)[:alpha] = Dict{Tuple{Int,Int},Union{JuMP.VariableRef,Int}}()
+
+    N = ids(pm, nw, :blocks)
+    L = ref(pm, nw, :block_pairs)
+
+    ir = ref(pm, nw, :substation_blocks)
+
+    for (i,j) in L
+        for k in filter(k->k∉ir,N)
+            var(pm, nw, :f)[(k, i, j)] = JuMP.@variable(pm.model, base_name="$(nw)_f_$((k,i,j))")
+            var(pm, nw, :f)[(k, j, i)] = JuMP.@variable(pm.model, base_name="$(nw)_f_$((k,j,i))")
+        end
+        var(pm, nw, :lambda)[(i,j)] = relax ? JuMP.@variable(pm.model, base_name="$(nw)_lambda_$((i,j))", lower_bound=0, upper_bound=1) : JuMP.@variable(pm.model, base_name="$(nw)_lambda_$((i,j))", binary=true)
+        var(pm, nw, :lambda)[(j,i)] = relax ? JuMP.@variable(pm.model, base_name="$(nw)_lambda_$((j,i))", lower_bound=0, upper_bound=1) : JuMP.@variable(pm.model, base_name="$(nw)_lambda_$((j,i))", binary=true)
+        var(pm, nw, :beta)[(i,j)] = relax ? JuMP.@variable(pm.model, base_name="$(nw)_beta_$((i,j))", lower_bound=0, upper_bound=1) : JuMP.@variable(pm.model, base_name="$(nw)_beta_$((i,j))", binary=true)
+    end
+    for (s,sw) in ref(pm, nw, :switch)
+        (i,j) = (ref(pm, nw, :bus_block_map, sw["f_bus"]), ref(pm, nw, :bus_block_map, sw["t_bus"]))
+        var(pm, nw, :alpha)[(i,j)] = var(pm, nw, :alpha)[(j,i)] = var(pm, nw, :switch_state, s)
+    end
+
+    f = var(pm, nw, :f)
+    λ = var(pm, nw, :lambda)
+    β = var(pm, nw, :beta)
+    α = var(pm, nw, :alpha)
+
+    for k in filter(kk->kk∉ir,N)
+        bp_to = filter(((j,i),)->i∈ir&&i!=j,L)
+        bp_fr = filter(((i,j),)->i∈ir&&i!=j,L)
+        if !(isempty(bp_fr) && isempty(bp_to))
+            c = JuMP.@constraint(
+                pm.model,
+                sum(f[(k,j,i)] for (j,i) in bp_to) -
+                sum(f[(k,i,j)] for (i,j) in bp_fr)
+                ==
+                -1.0
+            )
+        end
+
+        bp_to = filter(((j,i),)->i==k&&i!=j,L)
+        bp_fr = filter(((i,j),)->i==k&&i!=j,L)
+        if !(isempty(bp_fr) && isempty(bp_to))
+            c = JuMP.@constraint(
+                pm.model,
+                sum(f[(k,j,k)] for (j,i) in bp_to) -
+                sum(f[(k,k,j)] for (i,j) in bp_fr)
+                ==
+                1.0
+            )
+        end
+
+        for i in filter(kk->kk∉ir&&kk!=k,N)
+            bp_to = filter(((j,ii),)->ii==i&&ii!=j,L)
+            bp_fr = filter(((ii,j),)->ii==i&&ii!=j,L)
+            if !(isempty(bp_fr) && isempty(bp_to))
+                c = JuMP.@constraint(
+                    pm.model,
+                    sum(f[(k,j,i)] for (j,ii) in bp_to) -
+                    sum(f[(k,i,j)] for (ii,j) in bp_fr)
+                    ==
+                    0.0
+                )
+            end
+        end
+
+        for (i,j) in L
+            JuMP.@constraint(pm.model, f[(k,i,j)] >= 0)
+            JuMP.@constraint(pm.model, f[(k,i,j)] <= λ[(i,j)])
+            JuMP.@constraint(pm.model, f[(k,j,i)] >= 0)
+            JuMP.@constraint(pm.model, f[(k,j,i)] <= λ[(j,i)])
+        end
+    end
+
+    JuMP.@constraint(pm.model, sum((λ[(i,j)] + λ[(j,i)]) for (i,j) in L) == length(N) - 1)
+
+    for (i,j) in L
+        JuMP.@constraint(pm.model, λ[(i,j)] + λ[(j,i)] == β[(i,j)])
+        JuMP.@constraint(pm.model, α[(i,j)] <= β[(i,j)])
+    end
+end